refactor: rename BatchOperationInitializationHelper to BatchOperationInitializationBehavior
This follows the *Behavior naming convention used by other classes in the engine
(e.g., CatchEventBehavior, CommandDistributionBehavior, ExpressionBehavior)
