+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/batchoperation/scheduler/BatchOperationExecutionScheduler.java
@@ -58,7 +58,7 @@ import org.slf4j.LoggerFactory;
  * processing.
  *
  * @see StreamProcessorLifecycleAware
- * @see BatchOperationInitializationHelper
+ * @see BatchOperationInitializationBehavior
  * @see BatchOperationRetryHandler
  */
 public class BatchOperationExecutionScheduler implements StreamProcessorLifecycleAware {
@@ -66,7 +66,7 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
 
   private final Duration initialPollingInterval;
   private final BatchOperationState batchOperationState;
-  private final BatchOperationInitializationHelper batchOperationInitializer;
+  private final BatchOperationInitializationBehavior batchOperationInitializer;
   private final BatchOperationRetryHandler retryHandler;
   private final AtomicBoolean executing = new AtomicBoolean(false);
   private final AtomicReference<ExecutionLoopState> initializing = new AtomicReference<>();
@@ -75,7 +75,7 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
 
   public BatchOperationExecutionScheduler(
       final Supplier<ScheduledTaskState> scheduledTaskStateFactory,
-      final BatchOperationInitializationHelper batchOperationInitializer,
+      final BatchOperationInitializationBehavior batchOperationInitializer,
       final BatchOperationRetryHandler retryHandler,
       final Duration batchOperationSchedulerInterval) {
